pyproject.toml
[tool.poetry]
name = "Gamble-like Agent"
version = "0.1.0"
description = ""
authors = ["Bofan Gong <bfangong@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
cdp-sdk = "^0.0.7"
openai = "^1.52.2"
pytest = "^8.3.3"
flask = "^2.2.2"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"